+++ packages/components/fields/date-time-field/src/date-time-field.story.js
@@ -9,9 +9,9 @@ import {
   select,
   object,
 } from '@storybook/addon-knobs/react';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
-import * as icons from '../../icons';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
+import * as icons from '../../../icons';
 import DateTimeField from './date-time-field';
 
 storiesOf('Components|Fields', module)
