--- src/components/fields/date-time-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './date-time-field';
